+++ src/main/java/roomescape/member/service/AuthService.java
@@ -1,5 +1,9 @@
 package roomescape.member.service;
 
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 import org.springframework.stereotype.Service;
 import roomescape.member.controller.request.TokenLoginCreateRequest;
 import roomescape.member.controller.response.MemberResponse;
@@ -8,9 +12,12 @@ import roomescape.member.domain.Email;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Password;
 import roomescape.member.infrastructure.JwtTokenProvider;
+import roomescape.member.resolver.UnauthenticatedException;
 
 @Service
 public class AuthService {
+
+    private static final String TOKEN = "token";
     private final JwtTokenProvider jwtTokenProvider;
     private final MemberRepository memberRepository;
 
@@ -23,8 +30,8 @@ public class AuthService {
         Email email = new Email(tokenLoginCreateRequest.email());
         Password password = new Password(tokenLoginCreateRequest.password());
 
-        if (!memberRepository.isExistUser(email.getEmail(), password.getPassword())) {
-            throw new IllegalArgumentException("[ERROR] 등록되지 않은 회원입니다.");
+        if (!memberRepository.existsByEmailAndPassword(email, password)) {
+            throw new IllegalArgumentException("[ERROR] 아이디 또는 비밀번호를 올바르게 입력해주세요.");
         }
 
         String accessToken = jwtTokenProvider.createToken(tokenLoginCreateRequest.email());
@@ -33,8 +40,32 @@ public class AuthService {
 
     public MemberResponse findUserByToken(String token) {
         String payload = jwtTokenProvider.getPayload(token);
-        Member member = memberRepository.findUserByEmail(payload);
-        return new MemberResponse(member.getName(), member.getRole());
+        Optional<Member> member = memberRepository.findByEmail(new Email(payload));
+        if (member.isPresent()) {
+            return new MemberResponse(member.get().getName(), member.get().getRole());
+        }
+        throw new NoSuchElementException("[ERROR] 멤버가 존재하지 않습니다.");
     }
 
+    public String extractToken(HttpServletRequest request) {
+        Cookie[] cookies = request.getCookies();
+        if (cookies == null) {
+            throw new UnauthenticatedException("[ERROR] 로그인 정보가 유효하지 않습니다.");
+        }
+
+        String token = extractTokenFromCookie(cookies);
+        if (token == null) {
+            throw new UnauthenticatedException("[ERROR] 로그인 정보가 유효하지 않습니다.");
+        }
+        return token;
+    }
+
+    private String extractTokenFromCookie(Cookie[] cookies) {
+        for (Cookie cookie : cookies) {
+            if (TOKEN.equals(cookie.getName())) {
+                return cookie.getValue();
+            }
+        }
+        return null;
+    }
 }
